line circle and polygon circle.

+++ shared-module/vectorio/__init__.h
@@ -16,4 +16,4 @@ typedef struct {
 } vectorio_event_t;
 
 float measure_distance(
-    int x1, int y1, int x2, int y2);
+    float x1, float y1, float x2, float y2);
